requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.0.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
idna==3.4
inflection==0.5.1
jsonschema==4.16.0
mypy-extensions==0.4.3
oauthlib==3.2.1
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing-inspect==0.8.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
pydantic==1.10.2
pytest-django==4.5.2
pytest_factoryboy==2.5.1